# docs/API.yaml
openapi: 3.0.0
info:
  title: Virtual Bank API
  description: API for managing virtual bank accounts and transactions
  version: 1.0.0
servers:
  - url: https://virtualbank.server/v1
paths:
...
  /account/{id}:
    get:
      summary: Retrieve account by ID
      tags:
        - Accounts
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/AccountNotFound'
  /new_account:
    post:
      summary: Create a new account
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/InvalidRequest'
...
components:
  parameters:
    AccountId:
      name: id
      in: path
      description: ID of the account
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    AccountNotFound:
      description: Account not found
...
    InvalidRequest:
      description: Invalid request
...
  schemas:
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lastName:
          type: string
        firstName:
          type: string
        dateOfBirth:
          type: string
          format: date
        monthlyNetSalary:
          type: number
        uniqueAccountNumber:
          type: string
        overdraftStatus:
          type: boolean
        principalBalance:
          type: number
...
    AccountInput:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        dateOfBirth:
          type: string
          format: date
        monthlyNetSalary:
          type: number
...